feat: 전체 뉴스 조회 모델과 상세 뉴스 조회 모델 분리
