+++ iOS/RollTheDice/RollTheDice/Source/View/News/NewsCard/NewsListViewModel.swift
@@ -14,16 +14,16 @@ import SwiftUI
 // Macro -> iOS 17.0 이상부터. @Published 안 해도 됨.
 // 사용할 때 var 형식으로 해도 됨.
 @Observable class NewsListViewModel {
-    var newsList: [News]?
+    var newsList: [NewsList]?
+//    var selectedNewsId: Int?
     var currentPage = 0
     var newsCancellable: AnyCancellable?
-    
     
     
     //TODO: Plugin Settings
     let provider = MoyaProvider<NewsService>(plugins: [MoyaLoggingPlugin()])
     
-    func newsToViewModel(_ list: [News]) {
+    func newsToViewModel(_ list: [NewsList]) {
         self.newsList = list
     }
 }
@@ -47,7 +47,7 @@ extension NewsListViewModel {
         )
         .compactMap { $0.response?.data }
         .receive(on: DispatchQueue.main)
-        .decode(type: [News].self, decoder: JSONDecoder())
+        .decode(type: [NewsList].self, decoder: JSONDecoder())
         .sink(receiveCompletion: { result in
             switch result {
                 // success
